refactor: #870 중복 선언 분리
