+++ apps/web/src/app/desktop/component/retrospect/template/list/TemplateListItem/index.tsx
@@ -10,7 +10,7 @@ import { retrospectInitialState } from "@/store/retrospect/retrospectInitial";
 import { TemplateListConform } from "../TemplateListConform";
 import { useSearchParams } from "react-router-dom";
 import { useApiPostTemplateChoiceListView } from "@/hooks/api/backoffice/useApiPostTemplateChoiceListView";
-import { TemplateChoiceFormTag } from "@/types/template";
+import { resolveFormTag } from "@/utils/template/resolveFormTag";
 
 export type DesktopTemplateListItemProps = {
   id: number;
@@ -19,21 +19,6 @@ export type DesktopTemplateListItemProps = {
   imageUrl?: string;
   date?: string;
 };
-
-/**
- * 화면에 보여주는 템플릿 태그 문자열을 백엔드 통계 API의 enum 값으로 매핑합니다.
- * 매핑되지 않은 케이스는 커스텀 템플릿(CUSTOM)으로 간주합니다.
- */
-const TEMPLATE_TAG_TO_FORM_TAG: Record<string, TemplateChoiceFormTag> = {
-  KPT: "KPT",
-  "5F": "FIVE_F",
-  "Mad Sad Glad": "MAD_SAD_GLAD",
-  SSC: "SSC",
-  PMI: "PMI",
-  무제: "UNTITLED",
-};
-
-const resolveFormTag = (tag: string): TemplateChoiceFormTag => TEMPLATE_TAG_TO_FORM_TAG[tag] ?? "CUSTOM";
 
 export function TemplateListItem({ id, title, tag, imageUrl }: DesktopTemplateListItemProps) {
   const { openFunnelModal, closeFunnelModal } = useFunnelModal();
